Separated Hexagon and HabitBoard into separate components
